fixes primary IDs that were DECIMALs

+++ api/sqs-data-consumer/database/migrations/20230218001011-create-page-loads.js
@@ -7,7 +7,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       identifier: {
         type: Sequelize.STRING
